+++ lib/src/client/clients/browser_client.dart
@@ -5,5 +5,3 @@ import 'package:rest_client/rest_client.dart';
 http.Client createHttpClient({Proxy? proxy}) => http.Client();
 
 Future<dynamic> processJson(String body) => Future.value(json.decode(body));
-
-bool shouldSpawnIsolate() => false;